convert routes file to typescript, add events/jobs routes
