+++ src/pages/AllEventsPage.jsx
@@ -2,14 +2,13 @@ import React, { Component, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import EventCard from '../compoents/EventCard';
 import { Grid } from '@mui/material';
+import { routes } from '../lib/routes';
 
 export default function AllEventsPage({}) {
     const [ events, setEvents ] = useState([]);
 
     useEffect(() => {
-        const url = import.meta.env.VITE_APP_API_BASE_URL + 'event/';
-
-        fetch(url, {
+        fetch(routes.api.events.list(), {
             method: 'GET',
             headers: {
                 'Content-Type': 'application/json',
